pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["ilyaKrivitskiy <krivitskiy-ilya@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.3"
python-dotenv = "^1.0.0"
docker = "^6.1.3"
postgres = "^4.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.2"
generics = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
